Undefine `USEAZ` and restore `-quicklaunch`/`-autojoin` command line args
The FZ build configurations were still defining the `USEAZ` preprocessor definition, which they are probably not supposed to. This definition is currently only used for the introscreen. The non-USEAZ case was broken in one case, probably just accidentally by an additional line comment on the else case in 3d7bccb.

This also restores the functionality of the `-quicklaunch`/`-autojoin` command line arguments, primarily aimed at faster testing cycles during development.